refactor: rename useName hook file name
